feat: add types and Main component for main page
