+++ src/components/MainPage/Main.tsx
@@ -3,38 +3,14 @@
 import type { FC } from 'react';
 
 import { useAuth } from '@/hooks/useAuth';
+import type { MainMessages } from '@/types/main';
 
 import CallToAction from './CallToAction';
 import Hero from './Hero';
 import ProjectInfo from './ProjectInfo';
 import Team from './Team';
+import Thanx from './Thanx';
 import styles from './main.module.css';
-
-export interface TeamMember {
-  name: string;
-  role: string;
-  description: string;
-  url: string;
-}
-
-export interface TeamMessages {
-  title: string;
-  memberAlice: TeamMember;
-  memberBob: TeamMember;
-  memberCharlie: TeamMember;
-}
-
-export interface ProjectInfoMessages {
-  title: string;
-  description1: string;
-  description2: string;
-  description3: string;
-}
-
-export interface MainMessages {
-  team: TeamMessages;
-  projectInfo: ProjectInfoMessages;
-}
 
 interface MainProps {
   messages: MainMessages;
@@ -45,9 +21,12 @@ const Main: FC<MainProps> = ({ messages }) => {
 
   return (
     <div className={styles.mainWrapper}>
-      <Hero isAuth={isAuth} username={username} />
-      <CallToAction />
+      <Hero isAuth={isAuth} username={isAuth ? username : undefined} />
+
+      {!isAuth && <CallToAction />}
+
       <Team messages={messages.team} />
+      <Thanx />
       <ProjectInfo messages={messages.projectInfo} />
     </div>
   );
